index controller to CBV
