+++ products/views.py
@@ -1,16 +1,22 @@
 from django.shortcuts import render
-from .models import Product, ProductCategory
-
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.views.generic.base import TemplateView
+
+from common.views import CommonContextMixin
+from .models import Product, ProductCategory
 
 import json
 
 # Create your views here.
 
 
-def index(request):
-    context = {'title': 'geekShop'}
-    return render(request, 'products/index.html', context)
+class IndexView(CommonContextMixin, TemplateView):
+    template_name = 'products/index.html'
+    title = 'GeekShop'
+
+# def index(request):
+#     context = {'title': 'geekShop'}
+#     return render(request, 'products/index.html', context)
 
 
 def products(request, category_id=None, page=1):
